fix: goroutine leak and race condition in sherlock
The current implementation of Sherlock has checks for repetitive
Start & Stop, but it does not work all the time, for example:

```go
s.Start() // s.closed = 0 and start a goroutine
s.Stop() // s.closed = 1 but the goroutine may not be closed
s.Start() // s.closed = 0 and start a new goroutine
```

result in two goroutines for the dump loop, and there may
be data races between the two goroutines.

However, the checks to `s.closed` makes developers feel the code
is safe, and can handle repetitive Start & Stop correctly, which
introduces potential bugs to their code.

To be precise, this PR hasn't fix the issues directly, because it
is very very hard to make things always correct here. This PR
takes another approach, it removes the checks and leave all the
issues to the callers, where is a better place to handle them
correctly. For example, the sherlock service already leverage
`sync.Once` to avoid repetitive Start & Stop.

Signed-off-by: Bomin Zhang <localvar@hotmail.com>